requirements.txt
pygame
pypubsub
pyyaml
numpy
argparse
python-can
pyserial
at_serial_can
